Remove datagen checkstyle suppressions and refactor code.
